arch: cast page_size to unsigned long in PAGE_MASK
page_size is unsigned int, so ~(page_size - 1) produces a 32-bit mask
even on 64-bit targets. Cast to unsigned long so the complement extends
to the full pointer width.

+++ include/arch.h
@@ -66,7 +66,7 @@
 #define SYSCALL_OFFSET 0
 #endif
 
-#define PAGE_MASK (~(page_size - 1))
+#define PAGE_MASK (~((unsigned long)(page_size) - 1))
 extern unsigned int page_size;
 
 extern bool biarch;
